Update PDFConvert.cs.GetWkhtmlToPdfExeLocation()
Latest downloads from http://wkhtmltopdf.org/downloads.html are installing  by default the wkhtmltopdf.exe bits under c:/Program Files/wkhtmltopdf/bin folder. Both System Variables for Program Files are returning the same folder, usually with the (x86) suffix. This change checks if the location are equal and if it is so, forces the suffix out of the folder name and test the executable location. Tested in both english and portuguese windows versions.